rename transformer to processor
